fix delete functionality for Sprints add delete functionality for Userstory

+++ scrum/src/main/groovy/de/fh_zwickau/pti/geobe/service/SprintService.groovy
@@ -119,6 +119,12 @@ class SprintService {
     }
 
     public deleteSprint(SprintDto.CDelete command) {
-        sprintRepository.delete(command.id)
+        Sprint delete = sprintRepository.findOne(command.id)
+        if (delete) {
+            delete.backlog.all.toArray().each { Task it ->
+                delete.backlog.remove(it)
+            }
+            sprintRepository.delete(command.id)
+        }
     }
 }
